fixing population of profiles list in extension
I had to change the way the profiles list populates. It is easier now to
reload the list when the profiles change, that way everything is dynamic.
I had to make a new function for loading the profiles list, but also had
to ensure it was "clean" before adding the profiles to it.