rewriting logic for recursive directory loading
After having some issues with how the recursive directories are loaded I decided to rewirte
the logic controlling it. Now instead of making the load_profile() method setup things it
probably shouldn't, the monitors are setup on the _parse_info() section so all subdirectories
will still be monitored for changes.