fixing error in inital loading of profile
When the profile is loaded, the directories were not being watched unless they were a subdirectory
of the main directory specified. I've changed the load_profile() and _parse_info() methods to
better handle this. Now instead of trying to figure out if its a dir or image in the load_profile()
method, the _parse_info() method is called reguardless and it does the heavy lifting.