making daemon pay attention to interval
The preferences dialog now saves changes to the interval. When the interval changes, the daemon should pick up
on the change and cycle the timer so that it uses the new interval. It will not update the time remaining
because we do not track this, we just reset it back to 0.