This makes the basic switching work.
I had to make a minor update to settings, it should correctly return the
handler_id of the signal it just created since disconnect will correctly remove
it from the stack when called.

I also updated how the extension was initalized and how the settings for the
extension were used. Doing this, I've enabled the switching capabilites, but
nothing shows up yet. Next is actually populating the menu and making it all
show up.