Updating to work in hourly request from #3
This is the beginning of work to enable the hourly feature request from
BigE/desk-changer#3 With this change, the settings are updated to include a new
key called rotation. This is where the daemon and extension will read from for
now to enable the timer. Currently I've updated the timer to run every 5
seconds if hourly is selected. It will, however, only switch to the next
wallpaper at the begining of every hour.