Updating extension to add rotation setting
I updated the verbage of the rotation setting from timer to interval. I did
this because it changes the thought of the daemon from a timer to more of a
rotation behavior. This basically allows for future development to include
other rotation types specified by their own intervals during the day.

This commit updates the extetnsion itself to support the switch from a timer
to a rotation method. The only remaining thing to do is purge the timer-enabled
setting and possibly add the new rotation setting to the preferences for more
fine grain control.