Inital rewrite of how the daemon handles profiles
Profiles are now their own objects and self sustaining for all signals,
settings, etc, needed for each profile. This enables more control for each
profile as well as the ability to break things out more and give more control
to the daemon later.

The lockscreen profile is not complete. This is not a fully functional
change... yet.