Updating for more clarity in #29
If notifications are enabled, they will now be displayed when the rotation
setting or the random setting is changed. A simple message telling the user
how rotation will work will be displayed once the setting is updated. Same
when the random setting is changed. This will hopefully clear up some confusion
with the controls.