Fixing #50 crashing gnome-shell when interval is 0
I've updated the schema to set a range for the interval. Currently I've got
this set as 1 for the minimum and 86400 (1 day) for the maximum.

I've also updated the settings, that way if the user somehow sets a value
lower than 1, it will not actually store the value. This should stop things
from going too badly.