Working lockscreen profile & misc refactor
This update is a signifigant refactor in how the daemon and the profiles work
together. This update really separates the logic so the daemon is responsible
for rotation, while the profile is purely responsible for tracking its own
state.

The daemon itself now handles nearly all of the rotation related code. There is
a minor exception for the lockscreen, which only needs to rotate the wallpaper
when it is inherited from the desktop profile. The daemon now also manages
updating the lockscreen wallpaper no matter if it inherits the desktop or not.