Updating extension to make code more uniform
Updating properties to follow the mapping strategy that is laid out here:
https://gitlab.gnome.org/GNOME/gjs/-/blob/master/doc/Mapping.md#gobject-properties
This changes properties with an underscore (_) to use dashes (-) in the
ParamSpec naming.

The other change was to the destroy method of all objects, just making the code
similar in all of the objects we create in the extension.