Adding quick settings menu to extension
This addresses the issues in #80 to wrap up this issue. The extension itself
now has a quick settings menu that can be expanded to show some of the settings
that are already avaliable in the preferences. This helps keeps the UI from
being too cluttered while still giving access to these settings.