Updating Start/Stop to handle profile load/unload
Through testing it became apparent that a stop/start of the service would just
forward to the next wallpaper on start instead of continuing where it left off
when stopped. I've updated the code to reflect the unload in Stop and the load
inside of Start.

With this change I've also added the profile destroy() call inside of the
runner destroy call. Hopefully this will help solve the problem of the queue
not being saved on logout.