Updating timer classes, adding docs and new signal
This update adds a signal to the timer giving the callback multiple
opportunities to be called. The callback can still be passed into the timer
constructor as a property as well.

The constructor properties have been defined in a much more GObject way and
extend each other for easier use. This is reflected in all of the timers now.

I've updated the timers with as much documentation as possible to help show
how it all works together to execute the callback.